% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawls.R
\name{pawls}
\alias{pawls}
\title{Penalized adaptive weighted least squares regression}
\usage{
pawls(x, y, lambda1 = NULL, lambda2 = NULL, nlambda1 = 100,
  nlambda2 = 50, lambda1.min = 0.05, lambda2.min = 0.001, beta0 = NULL,
  w0 = NULL, initial = c("uniform", "PAWLS"), delta = 1e-06,
  maxIter = 1000, intercept = TRUE, standardize = TRUE,
  search = c("cross", "all"))
}
\arguments{
\item{x}{a numeric matrix containing the predictor variables without an intercept.  \code{pawls}
standardizes the data and includes an intercept by default.}

\item{y}{a numeric vector containing the response variable.}
}
\description{
Compute weighted least squares regression with \eqn{L_{1}}{L1} regularization on both the
coefficients and weight vectors.
}

